Revert "Add missing base_url for production"
This reverts commit d15baf71250f7cdde20374b1d9cb689c94f1cd93.

The misspelling in the next commit caused the environment variable value
from Digital Ocean to be not read by the app properly.

I had added the base_url in this file, but it wasn't needed if I had just
found and fixed the spelling in the next commit. So, reverting this commit.

See next commit.