Ch10: Encode Auth error as query parameter in case of failed login
The book suggests to use
```
    HttpResponse::build(self.status_code())
        .insert_header((
        LOCATION,
        format!("{}?error={}", login_route(), encoded_error),
    ))
    .finish()
```

But with the above, when a login fails, the url is not encoded with the
`encoded_error` as it should be.
Apparantly, modern browsers interpret this as an "Unauthorized" response,
meaning that the user does not have permission to access the resource and
so won't follow the redirect.

Just building the response like the following seems to work since
`SeeOther` explicitly tells the browser to redirect.
```
    HttpResponse::SeeOther()
        .insert_header((
        LOCATION,
        format!("{}?error={}", login_route(), encoded_error),
    ))
    .finish()
```